+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
+## Unreleased
+
+### Added
+
+- New commands:
+  - `code42 users deactivate`
+  - `code42 users reactivate`
+  - `code42 users bulk deactivate`
+  - `code42 users bulk reactivate`
+
 ## 1.9.0 - 2021-08-19
 
 ### Added